Move fixedxor implementation and test to set1 package
