--- set1/fixedxor.go
@@ -1,5 +1,0 @@
-package hextobase64
-
-func FixedXor(s1 string, s2 string) string {
-	return s1 + s2
-}
