+++ set1/set1.go
@@ -10,3 +10,13 @@ func Hextobase64(s string) string {
 	base64_result := base64.StdEncoding.EncodeToString(raw_bytes)
 	return base64_result
 }
+
+func FixedXor(s1 string, s2 string) string {
+	s1_bytes, _ := hex.DecodeString(s1)
+	s2_bytes, _ := hex.DecodeString(s2)
+	var xor_output []byte
+	for i := 0; i < len(s1_bytes); i++ {
+		xor_output = append(xor_output, s1_bytes[i]^s2_bytes[i])
+	}
+	return hex.EncodeToString(xor_output)
+}
